Refactor: create stress accessor base class.
